Tests/LibWeb/Text/input/HTML/document-open-from-deferred-script.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(async done => {
        window.deferredScriptRan = false;

        const iframe = document.createElement("iframe");
        const initialLoad = new Promise(resolve => iframe.addEventListener("load", resolve, { once: true }));
        document.body.appendChild(iframe);
        await initialLoad;

        const replacementLoad = new Promise(resolve => iframe.addEventListener("load", resolve, { once: true }));
        const iframeDocument = iframe.contentDocument;
        iframeDocument.open().write(`
            <script defer src="data:text/javascript,
                parent.deferredScriptRan%20%3D%20true%3B
                document.open()%3B
                document.write('%3Cbody%3EReplacement%20document%3C/body%3E')%3B
                document.close()%3B
            "><\/script>
            <body>First document</body>
        `);
        iframeDocument.close();

        await replacementLoad;
        println(`Deferred script ran: ${window.deferredScriptRan}`);
        println(`Replacement ready state: ${iframeDocument.readyState}`);
        println(`Replacement text: ${iframeDocument.body.textContent}`);
        iframe.remove();
        delete window.deferredScriptRan;
        done();
    });
</script>
