Tests/LibWeb/Text/expected/HTML/document-open-from-deferred-script.txt

Deferred script ran: true
Replacement ready state: complete
Replacement text: Replacement document
